wip1

+++ sliding_window_rf.py
@@ -10,7 +10,7 @@ from load_preproc_data import load_preproc_generator_windowed
 
 
 def make_rf_dataset(config):
-    return Dataset.from_generator(load_preproc_generator_windowed, (tf.float32, tf.uint32), (
+    return Dataset.from_generator(load_preproc_generator_windowed, (tf.float16, tf.uint8), (
         tf.TensorShape([config['window_size_y'] * config['window_size_x']]), tf.TensorShape([1])),
         args=(config['path_to_ubc3v'], config['window_size_x'], config['window_size_y']))
